requirements.txt
tensorflow-gpu>=2.3.0
scikit-learn
pandas
matplotlib
pyarrow>=0.17
sktime